interval type
